news_summarizer/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cssselect==1.2.0
# currentsapi==0.0.2
cymem==2.0.8
datasets==3.1.0
dill==0.3.8
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==2.8
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
langcodes==3.4.1
langdetect==1.0.9
language_data==1.2.0
lxml==5.3.0
lxml_html_clean==0.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.4.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.0.2
openai==1.54.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
# python-dateutil==2.8.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
# requests==2.21.0
requests-file==2.1.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
setuptools==75.5.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
sympy==1.14.0
textblob==0.18.0.post0
thinc==8.3.2
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.20.2
torch==2.7.1
tqdm==4.66.6
transformers==4.46.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.6
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.0
wikipedia==1.4.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.2
